requirements.txt
python_pushover==0.4
playsound==1.2.2
requests==2.25.1
discord_notify==1.0.0
pushbullet.py==0.9.1
infi.systray==0.1.12
plyer==2.0.0
infi==0.0.1
pushover==0.5
